Add delegate for implementing custom prompting behavior via the `ARKEmailBugReporterPromptingDelegate` protocol.
Also fixes a bug where the email dialog isn't shown when no log messages are included in the report. Previously, this could happen when the log stores are empty or the `emailAttachmentAdditionsDelegate` returns false for all log stores. The custom prompt adds another case of this bug by offering users the chance to filter out all log store.