Add convenience methods for logging to the distributor from Swift
This adds a couple of convenience methods for logging to a log distributor from Swift, to better support the use case of injecting the log distributor instead of using the global `log` function. The main advantage here over the Objective-C defined methods is the default method parameters and the better support for variadic Swift arguments.